Organize forms by options vs templates, rather than default vs POM
