+++ services/app-api/forms/measureTemplates.ts
@@ -3,7 +3,6 @@ import {
   ElementType,
   MeasureTemplateName,
   MeasurePageTemplate,
-  MeasureOptions,
 } from "../types/reports";
 import {
   additionalNotesField,
@@ -24,102 +23,10 @@ import {
   whichDeliverySystemsWereUsed,
 } from "./elements";
 
-export const defaultMeasures: MeasureOptions[] = [
-  {
-    cmit: 960,
-    uid: "960",
-    required: true,
-    stratified: false,
-    measureTemplate: [
-      MeasureTemplateName["LTSS-1"],
-      MeasureTemplateName["FFS-1"],
-      MeasureTemplateName["MLTSS"],
-    ],
-  },
-  {
-    cmit: 969,
-    uid: "969",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["FASI-1"]],
-  },
-  {
-    cmit: 961,
-    uid: "961",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-2"]],
-  },
-  {
-    cmit: 970,
-    uid: "970",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["FASI-2"]],
-  },
-  {
-    cmit: 20,
-    uid: "20",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-6"]],
-  },
-  {
-    cmit: 968,
-    uid: "968",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-7"]],
-  },
-  {
-    cmit: 414,
-    uid: "414",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-8"]],
-  },
-  {
-    cmit: 111,
-    uid: "111",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["HCBS-10"]],
-  },
-  {
-    cmit: 963,
-    uid: "963",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-3"]],
-  },
-  {
-    cmit: 962,
-    uid: "962",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-4"]],
-  },
-  {
-    cmit: 1255,
-    uid: "1255",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["LTSS-5"]],
-  },
-  {
-    cmit: 561,
-    uid: "561",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["MLTSS"]],
-  },
-];
-
-type TemplateMap = Partial<
-  Record<MeasureTemplateName, Omit<MeasurePageTemplate, "cmitId" | "status">>
->;
-
-export const defaultMeasureTemplates: TemplateMap = {
+export const measureTemplates: Record<
+  MeasureTemplateName,
+  Omit<MeasurePageTemplate, "cmitId" | "status">
+> = {
   [MeasureTemplateName["LTSS-1"]]: {
     id: "LTSS-1",
     title: "LTSS-1: Comprehensive Assessment and Update",
@@ -413,6 +320,147 @@ export const defaultMeasureTemplates: TemplateMap = {
       qualityMeasuresSubheader,
       qualityMeasureTable,
       additionalNotesField,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-1"]]: {
+    id: "POM-1",
+    title: "POM: People Live in Integrated Environments",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToRequiredDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      isTheStateReportingThisMeasure,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-2"]]: {
+    id: "POM-2",
+    title: "POM: People Participate in the Life of the Community",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToRequiredDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-3"]]: {
+    id: "POM-3",
+    title: "POM: People Choose Services",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToRequiredDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-4"]]: {
+    id: "POM-4",
+    title: "POM: People Realize Personal Goalss",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToRequiredDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-5"]]: {
+    id: "POM-5",
+    title: "POM: People are Free from Abuse and Neglect",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToRequiredDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-6"]]: {
+    id: "POM-6",
+    title: "POM: People Have the Best Possible Health",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToOptionalDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
+      measureFooter,
+    ],
+  },
+  [MeasureTemplateName["POM-7"]]: {
+    id: "POM-7",
+    title: "POM: People Interact with Other Members of the Community",
+    type: PageType.Measure,
+    sidebar: false,
+    elements: [
+      returnToOptionalDashboard,
+      measureHeader,
+      measureInstructions,
+      measureDetailsSection,
+      measureInformationSubheader,
+      didYouFollowSpecifications,
+      wereTheResultsAudited,
+      additionalNotesField,
+      whichDeliverySystemsWereUsed,
+      qualityMeasuresSubheader,
+      qualityMeasureTable,
       measureFooter,
     ],
   },
