--- services/app-api/forms/pomMeasures.ts
@@ -1,222 +1,0 @@
-import {
-  PageType,
-  MeasureTemplateName,
-  MeasurePageTemplate,
-  MeasureOptions,
-} from "../types/reports";
-import {
-  additionalNotesField,
-  didYouFollowSpecifications,
-  isTheStateReportingThisMeasure,
-  measureDetailsSection,
-  measureFooter,
-  measureHeader,
-  measureInformationSubheader,
-  measureInstructions,
-  qualityMeasuresSubheader,
-  qualityMeasureTable,
-  returnToOptionalDashboard,
-  returnToRequiredDashboard,
-  wereTheResultsAudited,
-  whichDeliverySystemsWereUsed,
-} from "./elements";
-
-export const pomMeasures: MeasureOptions[] = [
-  {
-    cmit: 1822,
-    uid: "1822-1",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-1"]],
-  },
-  {
-    cmit: 1822,
-    uid: "1822-2",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-2"]],
-  },
-  {
-    cmit: 1822,
-    uid: "1822-3",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-3"]],
-  },
-  {
-    cmit: 1822,
-    uid: "1822-4",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-4"]],
-  },
-  {
-    cmit: 1822,
-    uid: "1822-5",
-    required: true,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-5"]],
-  },
-  {
-    cmit: 1822,
-    uid: "1822-6",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-6"]],
-  },
-  {
-    cmit: 1822,
-    uid: "1822-7",
-    required: false,
-    stratified: false,
-    measureTemplate: [MeasureTemplateName["POM-7"]],
-  },
-];
-
-type TemplateMap = Partial<
-  Record<MeasureTemplateName, Omit<MeasurePageTemplate, "cmitId" | "status">>
->;
-
-export const pomMeasureTemplates: TemplateMap = {
-  [MeasureTemplateName["POM-1"]]: {
-    id: "POM-1",
-    title: "POM: People Live in Integrated Environments",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToRequiredDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      isTheStateReportingThisMeasure,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-  [MeasureTemplateName["POM-2"]]: {
-    id: "POM-2",
-    title: "POM: People Participate in the Life of the Community",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToRequiredDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-  [MeasureTemplateName["POM-3"]]: {
-    id: "POM-3",
-    title: "POM: People Choose Services",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToRequiredDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-  [MeasureTemplateName["POM-4"]]: {
-    id: "POM-4",
-    title: "POM: People Realize Personal Goalss",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToRequiredDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-  [MeasureTemplateName["POM-5"]]: {
-    id: "POM-5",
-    title: "POM: People are Free from Abuse and Neglect",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToRequiredDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-  [MeasureTemplateName["POM-6"]]: {
-    id: "POM-6",
-    title: "POM: People Have the Best Possible Health",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToOptionalDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-  [MeasureTemplateName["POM-7"]]: {
-    id: "POM-7",
-    title: "POM: People Interact with Other Members of the Community",
-    type: PageType.Measure,
-    sidebar: false,
-    elements: [
-      returnToOptionalDashboard,
-      measureHeader,
-      measureInstructions,
-      measureDetailsSection,
-      measureInformationSubheader,
-      didYouFollowSpecifications,
-      wereTheResultsAudited,
-      additionalNotesField,
-      whichDeliverySystemsWereUsed,
-      qualityMeasuresSubheader,
-      qualityMeasureTable,
-      measureFooter,
-    ],
-  },
-};
